+++ CHANGELOG.md
@@ -24,6 +24,7 @@
  - Fix Construct batches modifying more blocks than necessary (and hopefully fix issues with undo on large batches)
  - Fix hotbar spells getting lost on wand organize
  - Add "quiet" option to wands, useful for custom message display or for recording (quiet = 1 or 2 for less chat spam)
+ - Added "magic list maps <keyword>" command (magic list -> magic list wands)
 
 ## 2.9.7